+++ src/RestURL.ts
@@ -2,24 +2,26 @@ import type { TID } from "wx-lib-state";
 import { isTempID } from "wx-lib-state";
 
 function _addToUrl(url: string, ...parts: TID[]) {
+	if (!parts.length) return url;
+
 	parts.forEach(part => {
-		if (part) url += (url.endsWith("/") ? "" : "/") + encodeURIComponent(part);
+		if (typeof part === "undefined" || part === null || part === "") return;
+		const pstr = part.toString();
+		url +=
+			(url.endsWith("/") ? "" : "/") +
+			(pstr.startsWith("/") ? pstr.slice(1) : encodeURIComponent(pstr));
 	});
 
 	return url;
 }
 
-type RestURLConfig = {
-	headers?: Record<string, string>;
-};
-
 export class RestURL<T, R> {
 	private _url: string;
-	private _config: RestURLConfig;
+	private _headers: Record<string, string>;
 
-	constructor(url: string, config: RestURLConfig) {
+	constructor(url: string, headers?: Record<string, string>) {
 		this._url = url;
-		this._config = { headers: {}, ...(config || {}) };
+		this._headers = headers || {};
 	}
 
 	protected async fetch<X>(
@@ -42,17 +44,16 @@ export class RestURL<T, R> {
 		return result;
 	}
 
-	async get(id: TID): Promise<T> {
-		return this.fetch<T>(
-			_addToUrl(this._url, id),
-			"GET",
-			this._config.headers,
-			null
-		);
+	path(...parts: TID[]): RestURL<T, R> {
+		return new RestURL(_addToUrl(this._url, ...parts), this._headers);
 	}
 
-	async getAll(): Promise<T[]> {
-		return this.fetch<T[]>(this._url, "GET", this._config.headers, null);
+	headers(headers: Record<string, string>): RestURL<T, R> {
+		return new RestURL(this._url, { ...this._headers, ...headers });
+	}
+
+	async get(url?: TID): Promise<T> {
+		return this.fetch<T>(_addToUrl(this._url, url), "GET", this._headers, null);
 	}
 
 	async add(data: T): Promise<R> {
@@ -61,47 +62,47 @@ export class RestURL<T, R> {
 			"POST",
 			{
 				"Content-Type": "application/json",
-				...this._config.headers,
+				...this._headers,
 			},
 			data
 		);
 	}
 
-	async update(id: TID, data: T): Promise<R> {
-		if (isTempID(id)) return;
+	async update(data: T, url?: TID): Promise<R> {
+		if (url && isTempID(url)) return;
 
 		return this.fetch<R>(
-			_addToUrl(this._url, id),
+			_addToUrl(this._url, url),
 			"PUT",
 			{
 				"Content-Type": "application/json",
-				...this._config.headers,
+				...this._headers,
 			},
 			data
 		);
 	}
 
-	async delete(id: TID): Promise<R> {
-		if (isTempID(id)) return;
+	async delete(url?: TID): Promise<R> {
+		if (url && isTempID(url)) return;
 
 		return this.fetch<R>(
-			_addToUrl(this._url, id),
+			_addToUrl(this._url, url),
 			"DELETE",
-			this._config.headers,
+			this._headers,
 			null
 		);
 	}
 
-	async save(action: string, id: TID, data: T): Promise<R> {
+	async save(action: string, data: T, url?: TID): Promise<R> {
 		switch (action) {
 			case "add": {
 				return this.add(data);
 			}
 			case "update": {
-				return this.update(id, data);
+				return this.update(data, url);
 			}
 			case "delete": {
-				return this.delete(id);
+				return this.delete(url);
 			}
 		}
 	}