+++ src/index.ts
@@ -2,7 +2,7 @@ import Rest from "./RestDataProvider";
 import DataProvider from "./DataProvider";
 import ActionQueue from "./ActionQueue";
 
-export type { ActionMap } from "./types";
+export type { ActionMap, RestDataProviderConfig } from "./types";
 export { RestURL } from "./RestURL";
 export { Rest, ActionQueue, DataProvider };
 export { SYNC, NOTSENT, SENDING } from "./ActionQueue";